overlays: Rename gpio-fsm property num-soft-gpios
commit  d7c8eb8181c9892a9b9a0365de15b7e725d02d12 from
https://github.com/raspberrypi/linux.git rpi-5.10.y

The gpio-fsm property "num-soft-gpios" triggers a kernel DT checker
that warns about the lack of #gpio-cells on a random node with the
phandle that just happens to match the number of soft GPIOs. Rename
the property to "num-swgpios" to avoid the warning.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>