octeontx2-af: debugfs: delete dead code
commit 0f4de2a22e63e1c65eb289cd351373f2a1fda784 from
git@git.assembla.com:cavium/WindRiver.linux.git

These debugfs never return NULL so all this code will never be run.

In the normal case, (and in this case particularly), the debugfs
functions are not supposed to be checked for errors so all this error
checking code can be safely deleted.

This is a backported upstream patch with more changes for SSO etc.
Upstream patch
2ce5a30711a ("octeontx2-af: debugfs: delete dead code")

Change-Id: I885a5640463d49a11869dda38ba09c277f857e05
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/47859
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>