media: cadence: csi2rx: Turn subdev power on before starting stream
commit 86688e9f8a2e6912a005d9201ff28a2273669fab from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The subdevice power needs to be turned on before the stream is started.
Otherwise it might not be in the proper state to stream the data.

For some reason, it is observer with OV5640 that turning the power off
when stopping the stream causes the next stream to freeze infinitely. So
leave it on for now until this issue is root caused.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>