HACK: media: ti-vpe: csi2rx: Drain DMA when stopping stream
commit 3abf65738c67f85c2cecec403bb65e2ba5043304 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Some data might be stuck in the DMA pipeline because the application
does not tell us how many frames it wants to capture. So there will
always be some time delay between the application requesting the last
frame it needs and stopping the stream which will stop DMA. Drain that
data so it does not corrupt the next frame captured when the stream is
re-started later.

Marking this as a hack for now because it is not clear yet whether this
is a hardware problem or a software problem. If it does turn out to be a
hardware problem, it can be presented as a workaround instead.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>