driver: dma: fsl-edma: Check the HRS before stopping the DMA service request
We have observed a serial malfunction when using DMA on
a NXP s32g board. The reason is that the serial TX DMA doesn't
work if we stop the serial TX DMA service request dynamically.
According to the eDMA manual, there are certain restrictions to
disable a DMA hardware service request. For coherence, it must
follow a specific sequence like below:
1. Stop the DMA service request at the peripheral first.
2. Check the DMA's Hardware Request Status (HRS) to ensure
there is no service request to the DMA channel being stopped.
Then disable the hardware service quest by clearing the ERQ
field to 0 on the appropriate DMA channel.

So add the code to check the HRS before stopping the DMA service request.

Signed-off-by: Meng Li <Meng.Li@windriver.com>