pcie:s32gen1: Remove SoC version check when enabling PCIe
commit e3009a41784c8c52b2b14462aa8773057507a3d2 from
https://source.codeaurora.org/external/autobsps32/linux

This check is only valid for rev 1.0.0 S32G274A SoCs, which had
PCIe issues.
This revision is now obsolete and only existed on some of the
early S32G274AEVBs (and which did not get outside NXP).
More, the check is still performed in u-boot, which is enough.

Issue: ALB-4529
Issue: ALB-5981
Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>