arm64: dts: ti: k3-am64-main: Add hwspinlock node
commit cb1c192d14345d320a17b50d7101f0ac45d4e025 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The AM64x SoC contains a HwSpinlock IP instance that is a minor variant
of the IP on existing TI K3 SoCs such as AM65x, J721E or J7200 SoCs.
Add the DT node for this on AM64x SoCs. The node is present within the
MAIN domain, and is added as a child node under the cbass_main node.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>