misc: bphy: Add ioctl to get BPHY irqs bitmask
commit 9fbb29694b78d150ba43e5a935f09955962d2dd9 from
git@git.assembla.com:cavium/WindRiver.linux.git

Availability of BPHY interrupts varies between platforms.
This patch aims at providing information about supported
interrupts numbers to userspace. The information reported
to user is a bitmask with up to 64 where each bit position
corresponds to interrupt number. If bit is set this means
that interrupt is available. For example value 0x0A,
or 00001010b, indicates that irq numbers 1 and 3 are
available.
Currently the information of supported interrupts is held
by a 64 bit variable which sets an upper limit of 64 irqs.

Change-Id: I34864c8803b0a2dadac2b410ea40801d1645ec08
Signed-off-by: Jakub Palider <jpalider@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/47667
Reviewed-by: Radha Chintakuntla <radhac@marvell.com>
Reviewed-by: Naveen Mamindlapalli <Naveen.Mamindlapalli@cavium.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/56795
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>