net: xilinx: Handle TX TS FIFO timeout error
commit  ed8a776f2db433a413d29b2d86e2573731524269 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

When TX TS FIFO timeout error is observed, the FIFO should not be
popped as there is no data present. Handle the error case by skipping
the same.

This error condition may be observed on a malfunctioning design
where the HW does not provide a timestamp even when requested.
In such a case, if an empty FIFO is read, it will result in a crash.
Handling this error will ensure proper exit so that linux prompt
is present and further debug is possible.

Signed-off-by: Harini Katakam <harini.katakam@xilinx.com>
Signed-off-by: Sarath Babu Naidu Gaddam <sarath.babu.naidu.gaddam@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>