ARM: dts: am57xx: Add PRUSS MDIO controller nodes
commit d9fd486264ae2bb2a0cbab94bfd7333a2bf7ee68 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The PRUSSs on AM57xx SoCs contain an MDIO controller that can
be used to control external PHYs associated with the Industrial
Ethernet peripherals within each PRUSS. The MDIO module used
within the PRU-ICSS is an instance of the MDIO Controller used
in TI Davinci SoCs. The same bus frequency of 1 MHz is chosen as
the regular MDIO node.

The nodes are added to the common am57-pruss.dtsi file and are
disabled. These need to be enabled in the respective board files
supporting AM57xx SoCs and where the ethernet is pinned out and
connected properly.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>