arm64: dts: ti: k3-am642-evm: Add mailboxes to R5Fs
commit 11a1bcac5b99732eb824bdc832d847d37379c78b from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add the required 'mboxes' property to all the R5F processors for the
TI AM642 EVM board. The mailboxes and some shared memory are required
for running the Remote Processor Messaging (RPMsg) stack between the
host processor and each of the R5Fs.

The chosen sub-mailboxes match the values used in the current firmware
images. This can be changed, if needed, as per the system integration
needs after making appropriate changes on the firmware side as well.

Note that any R5F Core1 resources are needed and used only when that
R5F cluster is configured for Split-mode.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>