mtd: spi-nor: core: perform a Soft Reset on shutdown
commit eac49a70c13ad126ebbff0b54803f65d37df90f8 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit d73ee7534cc537b98b61bfd83dbce5bb12aecb80 upstream.

Perform a Soft Reset on shutdown on flashes that support it so that the
flash can be reset to its initial state and any configurations made by
spi-nor (given that they're only done in volatile registers) will be
reset. This will hand back the flash in pristine state for any further
operations on it.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Reviewed-by: Tudor Ambarus <tudor.ambarus@microchip.com>
Link: https://lore.kernel.org/r/20201005153138.6437-13-p.yadav@ti.com
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>