bus: fsl-mc: force device rescan if endpoint not discovered
commit 64cee2378a260a592c8bb7bfa58e75178f45c77f from
git://source.codeaurora.org/external/qoriq/qoriq-components/linux

If the endpoint of a device is not yet probed on the bus force
a rescan of the devices and retry to get a reference to the
endpoint device. If the device is still not found then we assume
it's in a different isolation context (container/DPRC) thus
unavailable and return a permission error.

Signed-off-by: Laurentiu Tudor <laurentiu.tudor@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>