octeontx2-pf: Fix bandwidth profile NIX_AQ_INSTOP_WRITE
commit 850e8e19ee021852174c353eb4732a19e6eeb490 from
git@git.assembla.com:cavium/WindRiver.linux.git

NIX admin quueue operation NIX_AQ_INSTOP_WRITE
performs a masked write to a NIX-managed hardware
structure. Software provides the data to be written
to the structure. Software must also provide a write
mask of the structure, where each bit to be updated
is a 1, and each bit not to be modified is a 0.
For configuring NIX ingress bandwidth profile context
the masks are missed. This patch fixes that.

Fixes: bd825b111cf3 ("octeontx2-pf: TC_MATCHALL ingress")
Change-Id: I640660e48b98902fe96101f852fde0a4402159fa
Signed-off-by: Subbaraya Sundeep <sbhatta@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/48932
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>