iio: adc: versal-sysmon: add status reset after temp alarm is handled
commit  3b9b181f3740c2240885438b9874b3b0b9f8626f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

When the temperature events are handled and the temperature values is in
the range of threshold, reset status is updated. This will enable user
to see the new max_max and min_min values. This is necessary in case
there is a new temperature event, the new min_min and max_max values
will enable user to identify if the event was due to crossing upper or
lower threshold.

Signed-off-by: Anand Ashok Dumbre <anand.ashok.dumbre@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>