trace/rt: fixup -stable backport
v5.12-rc6 got the following tracing change:

  9deb193af69d3f [tracing: Fix stack trace event size]

And it was very quickly (magically) backported to 5.10.28,
but was not a cherry pick, since the tracing flags are
different in the 5.10 series.

BUT, in 5.12 more of the preempt-rt tracing changes also
merged 36590c50b2d07 [tracing: Merge irqflags + preempt counter],
which means that the original 5.12 commit is actually correct
for 5.10-rt, not the ported -stable variant.

We substitute trace_ctx for the pc,flags seen in the 5.10
port, and -rt builds again.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>