LF-3401 spi: imx: remove CLK calculation and divider in slave mode
commit a216dbe15a896447bf5effffa43d5635b3545619 from
git://source.codeaurora.org/external/qoriq/qoriq-components/linux

In the slave mode, the clock signal is controlled by the master.
The slave does not need to calculate and configure the clock
frequency division. The slave can directly use the root clock
to sample the SCL signal.
Therefore, remove the calculation and frequency division
function of the clock in the slave mode.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>