overlays: Add ugreen-dabboard overlay
commit  73cee4f50858d1cb3834fa11d2bc7a1b4b6fe616 from
https://github.com/raspberrypi/linux.git rpi-5.10.y

This is a simple overlay based on the simple-audio-card and the dmic
codec. It has the speciality that it is configured to use the codec
as a master I2S device. It works for example with the Si468x DAB
receiver on the uGreen DABBoard.

See: https://github.com/raspberrypi/linux/issues/4304

Signed-off-by: Christoph Orth <c.orth@ugreen.eu>
Signed-off-by: Meng Li <Meng.Li@windriver.com>