dpaa2-eth: Set congestion group before the Rx queues become operational
commit cf8c29c14df7f7cacfe04b71e682cfe409f7ec22 from
git://source.codeaurora.org/external/qoriq/qoriq-components/linux

Setting  the CGID for the first time or dynamically changing it
with a different one  on a FQ that has been in service (has frames in it)
is not supported, the byte/frame counts in both the old and new CGRs may be
corrupted if this is done.

This patch adds the configuration of the CGR(congestion group record)
for the network interface RX queues, in the probe function before the queues
become operational,  to prevent the corruption of the aforementioned counter.

Signed-off-by: Radu Bulie <radu-andrei.bulie@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>