MLK-23277: 8qm: Fix SW workaround for i.MX8QM TKT340553
commit 4069a4ca1ea75bc732f552d3b11fb3574652dc65 from
git://source.codeaurora.org/external/qoriq/qoriq-components/linux

Current workaround is looping uselessly on the address
range doing a  _tlbi(vmalle1is) which is harmful for
the system performance and buggy as the instruction is
flushing the entire TLB and there is no benefit of
redoing it more than once. Also fix missing barriers.

Signed-off-by: Nitin Garg <nitin.garg@nxp.com>
Signed-off-by: Marouen Ghodhbane <marouen.ghodhbane@nxp.com>
Reviewed-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>