crypto: sa2ul - Add support for AM64
commit 791197e2dc8d43f8e05ab403443cdeede1588fd6 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 0bc42311cdff4ee8bdf08f68b28b0307895df90b upstream.

The sa2ul module in am64 have limited support for algorithms, and the
priv and priv_id used on the platform is different compared to AM654 or
j721e.

Use match data to get the SoC specific information and use it throughout
the driver.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Vaibhav Gupta <v_gupta@ti.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
[s-anna@ti.com: cherry-pick commit '0bc42311cdff' from v5.13]
Signed-off-by: Suman Anna <s-anna@ti.com>
Reviewed-by: Tero Kristo <kristo@kernel.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>