rtc:s32gen1: Save time before system suspend
commit ee5e656e996bbd847608c879ae2d531c6adb7dae from
https://source.codeaurora.org/external/autobsps32/linux

After suspend/resume we reinitialize the RTC, which causes a subsequent
rtc_read_time() to fail because the hardware counter is now behind what
we had saved in priv->base.
Save the time just before suspend and reapply at resume to fix hwclock
in combination with suspend-to-ram.
Also take into account the amount of time (in seconds) that we are due
to sleep so that it gets reflected on resume.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@nxp.com>
Issue: ALB-7084
Issue: ALB-6347
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>