ARM: dts: AM33xx/AM43xx: Rename wkup_m3 sub-mailbox node
commit 9f6685e37c1c393a46bd484b75a1ee4ddf7397d1 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The OMAP sub-mailbox used to communicate with the Wakeup M3 remote
processor is currently named wkup_m3. This name can be confused with
the remote processor node. So, rename this to mbox-wkup-m3 to remove
the ambiguity and to also adhere to the sub-mailbox node name convention
being added in the OMAP Mailbox YAML binding.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>
[s-anna@ti.com: cherry-pick linux-next commit '8e880dfefd61' for v5.14]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>