irqchip: xilinx: Add support to remove the Xilinx INTC driver module
commit  a6476c8a58b5a95109e5b9bb480efc775799ab07 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

The existing Xilinx INTC driver does not have mechanism to remove the
driver when it is being used as a module. For example, if the Xilinx
INTC IP is part of a removable fpga partition and a user intends to
replace the removable partition with a new one, the existing
implementation will not support the same.
The enhancement is done by adding a new remove API and making other
necessary changes. The added API xilinx_intc_of_remove ensures that all
interrupts already mapped to the interrupt domain and the corresponding
irq descriptors freed up.
Changes in this patch is made through a new irqchip kconfig option with the
name IRQCHIP_XILINX_INTC_MODULE_SUPPORT_EXPERIMENTAL which is disabled
by default.

Signed-off-by: Anirudha Sarangi <anirudha.sarangi@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>