drm: Introduce new state accessors in place of drm_atomic_get_crtc_state
commit  040d0dd82dbfad11984b696e35beb56906f8f4a7 from
https://github.com/raspberrypi/linux.git rpi-5.10.y

drm_atomic_get_crtc_state relies on drm_atomic_get_existing_crtc_state
that is deprecated and isn't really clear on which state it provides.

Even worse, after the states have been swapped, if the CRTC is present
in the state drm_atomic_get_existing_crtc_state will return the old
state, whereas if the CRTC is not present, we'll use a copy of
crtc->state. crtc->state for the CRTC that is there at that point is the new
state though, which leads to confusion.

Let's provide two new helpers that make it clear what our expectations
are.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>