sdk_dpaa: rework private bpool seeding per CPU per interface
commit ef3f2cfc6010c13feb40cfb7fd7490832cf86f45 from
git://source.codeaurora.org/external/qoriq/qoriq-components/linux

The dpa_bp introduction has modified the existing design of buffer
pool seeding. The most notable (and unintended) change was that
instead of placing in the buffer pool an allotment of buffers for
each CPU, for each interface, only an allotment for each CPU was
performed. With this, the possibility that in certain scenarios,
each interface can draw from the buffer pool enough buffers to
drain it completely, but not enough to trigger a refill. This may
result in a complete stall of the reception on the interface.

The current patch reintroduces the initial design, that makes sure
each probed interface contributes to the common buffer pool an
allotment of buffers for each CPU, removing the above mentioned
issue.

In consequence, the per buffer pool seed_cb() callback is removed.
The seeding is now done at net device probe.

Reported-by: Radu Bulie <radu-andrei.bulie@nxp.com>
Signed-off-by: Madalin Bucur <madalin.bucur@nxp.com>
Signed-off-by: Camelia Groza <camelia.groza@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>