xen/pvcalls: frontend enable
commit  8adabe58bb2d115b1a95bd97fdd51024d87ebcca from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

The Xen pvcalls frontend driver needs to be explicitly enabled, and
there isn't a way to do it upstream yet. This patch adds a simple
boolean command line option (gated on CONFIG_XEN_PVCALLS_FRONTEND) to
enable pvcalls. It is a temporary solution but it is the best we can do
until a proper upstream mechanism to enable pvcalls is introduced.

The patch has no effect if CONFIG_XEN_PVCALLS_FRONTEND is not enabled,
or if pvcalls=true is not passed as a command line argument.

Signed-off-by: Stefano Stabellini <stefano.stabellini@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>