drm/vc4: hdmi: Drop devm interrupt handler for CEC interrupts
commit  e6fa19ad1503290bbebe821e4c18c9b0e5666f0b from
https://github.com/raspberrypi/linux.git rpi-5.10.y

The CEC interrupt handlers are registered through the
devm_request_threaded_irq function. However, while free_irq is indeed
called properly when the device is unbound or bind fails, it's called
after unbind or bind is done.

In our particular case, it means that on failure it creates a window
where our interrupt handler can be called, but we're freeing every
resource (CEC adapter, DRM objects, etc.) it might need.

In order to address this, let's switch to the non-devm variant to
control better when the handler will be unregistered and allow us to
make it safe.

Fixes: 15b4511a4af6 ("drm/vc4: add HDMI CEC support")
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>