net: hsr: Save timestamp and tag info of stripped skb
commit ffd978f0107ce61348b25bee3a8d82006fd150f1 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Before the tag is stripped from a received skb, the tag info and
receive timestamp is saved in the shared info area of the skb so
that upper core layer can pass the saved info to user application.

In this patch the support of saving tag info and timestamp is added for
HSR PTP packets over L2.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>