arm64: dts: ti: k3-j721e-eaik: Add overlay for CSI2 interfaced with OV5640 camera
The OV5640 camera sensor is interfaced with CSI2RX through a MIPI header.
The sensor shall be configured through the I2C which is interfaced with
the camera sensor via an I2C switch.

MAIN_I2C ===> I2C_switch(TCA954) ====> \\
                                         ==>MIPI header ===> OV5640
                         CSI2RX0 ====> //

Add node for the I2C switch through which the camera sensor is connected.

Add node for OV5640 camera connected to port 0 of the CSI bridge.

Add pinmux for the reset gpio and CSI2 Reference clock to the camera
sensor.

Signed-off-by: Sinthu Raja <sinthu.raja@ti.com>
[Xulin: Original patch taken from TI PROCESSOR-SDK-LINUX-J721E_08.00.00.08]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>