media: ti-vpe: cal: Create subdev for CAMERARX
commit 01d4f990d859d27eee66b5c738246f64c879102c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 5acc3e22c25359f41ae18d3f8c63e20b6e4f74a8 upstream.

Create and register V4L2 subdevs for the CAMERARX instances, and link
them in the media graph to the sensors and video nodes. The subdev API
is not exposed to userspace at this point, and no subdev operation is
implemented, but the media controller graph is visible to applications.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>