media: ti-vpe: cal: rename "sensor" to "source"
commit 346e423a6bfd8ed32d577bb46332f026f195e235 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

CAL driver uses "sensor" name to refer to the subdev connected to CAL.
As the subdev can also be a bridge, the naming is misleading and might
cause the reader to think it refers to the actual sensor at the end of
the pipeline.

Rename "sensor" to "source".

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>