v4l: xilinx: Added support for DP audio
commit  b685798b681b67aaab3f532e03b256f60fc32de9 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

This patch is to add the support for DP RX audio capture. Implementation
of this patch is same as the hdmi audio driver. This patch registers the
dp audio handlers with the formatter driver and sound card driver to create
a audio sound card.This was tested with alsa framework (arecord).

Signed-off-by: Jagadeesh Banisetti <jagadeesh.banisetti@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>