drm: xlnx: dptx: Added support for DP audio
commit  b84c5749c24b28041593502b345c1f8d1782626b from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

This patch is to add the support for DP TX audio playback. Implementation
of this patch is same as the hdmi audio driver. This patch registers the
dp audio handlers with the formatter driver and sound card driver to create
a audio sound card.This was tested with alsa framework (aplay).

Signed-off-by: Jagadeesh Banisetti <jagadeesh.banisetti@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>