arm: dts: add am335x-boneblack-pps.dts
commit 6c83723bb3af683dc300f875461936de26d305fa from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add a new am335x-boneblack-pps.dts that can be used to configure TIMER7
in PWM mode. This PWM signal can be used as a PPS signal when
synchronized to PTP clock. Typically this synchronization is done by a
userspace program. This PPS signal can be observed on pin P8.8

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Adrien Ricciardi <aricciardi@baylibre.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>