llce: Replace memory pools with mapped areas
commit a93668920a9e1a927d4570ce4a0a222570f66e62 from
https://source.codeaurora.org/external/autobsps32/linux

This commit replaces the LLCE SRAM memory pools with
memory mapped areas in order to perform SRAM initialization
when it's needed only. The previous implementation delegated this
operation to SRAM driver and, as a direct consequence, the memory
was initialized even in cases where the LLCE management was split
between Linux and another software stack.

Issue: ALB-5830
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>