misc: xilinx-ai-engine: resolve refcounting issue
commit  fc89046e6a152fdb606b28172db8676bb376a5f8 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

When exiting the driver, the refcount is incorrect for both tile
and partition devices. This results in an issue where the release
functions were not firing and device managed resources were not
freeing. An example of the managed resource which did not free
was the IRQ. This prevented the kernel from reinserting the driver
successfully.

This patch resolves the refcounting issue and all devices reach 0
on driver exit.

Signed-off-by: Daniel Steger <daniel.steger@xilinx.com>
Reviewed-by: Wendy Liang <wendy.liang@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>