octeontx2-af: Re-apply: fix order of queue work and interrupt disable
commit 4fa9a19a144afffa69e34d9fc99c42bb8db41dae from
git@git.assembla.com:cavium/WindRiver.linux.git

Clear and disable interrupt before queueing work as there might be
a chance that work gets completed on other core faster and
interrupt enable as a part of the work completes before
interrupt disable in the interrupt context. This leads to
permanent disable of interrupt.

Re-applying this fix as it got reverted with below commit.

Fixes: I7208624243f "octeontx2-af: Update driver to keep sync with upstream driver"
Signed-off-by: Geetha sowjanya <gakula@marvell.com>
Change-Id: I05235a6a6786a4a39ab97605423bc4cb1dbf0703
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/58517
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>