driver: of: remove the redundant OF_CONFIGFS item in Kconfig
Recently, there is a patch that adds OF_CONFIGFS in drivers/of/Kconfig
in v5.10/standard/base branch, and it is merged into raspberrypi branch.
This causes build issue, because in raspberrypi branch, there has been
OF_CONFIGFS item. So, remove the redundant new OF_CONFIGFS, and
keep the raspberrypi own one to avoid conflicting.

Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>