clk: zynqmp: Enable the driver if ZYNQMP_FIRMWARE is selected
commit  eda08233e1a47a97256a830fcebf7b974016ccd6 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

When booting the kernel on zynqmp based platforms such as Ultra96v2,
peripheral drivers such as that for the sdcard depend on the presence
of clocks.

Enable the clock driver if it's dependencies are compiled to avoid
building an unbootable kernel.

Signed-off-by: Punit Agrawal <punit1.agrawal@toshiba.co.jp>
Link: https://lore.kernel.org/r/20210322061754.1065367-2-punit1.agrawal@toshiba.co.jp
Acked-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Stephen Boyd <sboyd@kernel.org>
Signed-off-by: Rajan Vaja <rajan.vaja@xilinx.com>
State: upstream ()
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>