Revert "drm: xlnx: vtc: defer owned VTC bridge un-registration"
commit  09a4215f61493584f4410cab3477003e4046dab9 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

struct xlnx_bridge is embedded in a container struct which may not valid
when the container driver is removed. Instead of delaying the
unregistering of the bridge in case it is still owned by other
component, do the unregistering it when the container is removed.

In addition,  change the WARN_ON dump to message print out to show the
case that the bridge is still owned by other components

Signed-off-by: Jianqiang Chen <jianqiang.chen@xilinx.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>