spi: spi-cadence-quadspi: Limit RX max tap to 127
commit  a9fe7944deb57c795851ed6abab57da988917728 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

RX tuning algorithm sweep for 128taps which is incorrect because in
controller, provision for configuring RX taps is only till 127taps.
Due to this, when SW driver tries to program 128th tap, it is becoming
zero and it is considered as Maximum tap, resultantly window size is
becoming zero. RX Tuning algorithm is failing with not meeting the minimum
required valid window size. Based on the discussion with design team,
restrict the RX max tap value to 127 and this patch does the same.

Signed-off-by: Sai Krishna Potthuri <lakshmi.sai.krishna.potthuri@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>