drm: xlnx: mixer: Use source rectangle for scaling check
commit  61670a725f57139f46c0b38e531b6fbde68d3f59 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

Use source rectangle  provided by application instead of
relying on framebuffer resolution to check if scaling is intended
or not.

This because if we are doing cropping than cropped resolution
may not match with framebuffer resolution but this would still work
fine if destination rectangle is also cropped and
same as source rectangle as there is no scaling involved
in this scenario.

Signed-off-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>