regulator: rpi-panel: Serialise operations.
commit  6ace7cd753d073081b38beb52422deb02aab71ad from
https://github.com/raspberrypi/linux.git rpi-5.10.y

The driver was using the regmap lock to serialise the
individual accesses, but we really need to protect the
timings of enabling the regulators, including any communication
with the Atmel.

Use a mutex within the driver to control overall accesses to
the Atmel, instead of the regmap lock.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>