regulator: rpi-panel: Add GPIO control for panel and touch resets
commit  2d8f0b900f1d5a7ceb4063936faa486374c6fd3e from
https://github.com/raspberrypi/linux.git rpi-5.10.y

We need independent control of the resets for the panel&bridge,
vs the touch controller.

Expose the reset lines that are on the Atmel's port C via the GPIO
API so that they can be controlled appropriately.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>