gpio-fsm: Clamp the delay time to zero
commit  597cd63b48c6dc3746827504f191744c120fbb4d from
https://github.com/raspberrypi/linux.git rpi-5.10.y

The sysfs delay_ms value is calculated live, and it is possible for
the time left to appear to be negative briefly if the timer handling
hasn't completed. Ensure the displayed value never goes below zero,
for the sake of appearances.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>